+++ util/geom.py
@@ -1,0 +1,31 @@
+
+def binom(n, k):
+	"""	
+	Parameters:
+		n - Number of elements of the entire set
+		k - Number of elements in the subset
+	It should hold that 0 <= k <= n
+	Returns - The binomial coefficient n choose k that represents the number of ways of picking k unordered outcomes from n possibilities
+	"""
+	answer = 1
+	for i in range(1, min(k, n - k) + 1):
+		answer = answer * (n + 1 - i) / i
+	return int(answer)
+
+def multivariate_hypgeom(deck, needed):
+	"""	
+	Parameters:
+		deck - A dictionary of cardname : number of copies
+		needed - A dictionary of cardname : number of copies
+	It should hold that the cardname keys of deck and needed are identical
+	Returns - the multivariate hypergeometric probability of drawing exactly the cards in 'needed' from 'deck' when drawing without replacement 
+	"""
+	answer = 1
+	sum_deck = 0
+	sum_needed = 0
+	for card in deck.keys():
+		answer *= binom(deck[card], needed.get(card, 0))
+		sum_deck += deck[card]
+		sum_needed += needed.get(card, 0)
+	return answer / binom(sum_deck, sum_needed)
+
